(chore) Make each app have independently runnable vitest tests
Mirrors the pattern Ian introduced for openmrs-esm-patient-chart in
PR #3301: a single shared vitest config in tools/, a thin re-export at
the workspace root, and a thin merge shim per app. With this in place,
'cd packages/<app> && yarn test' works without bouncing through the
workspace root.

- tools/vitest.shared.ts: the actual config that used to live at
  vitest.config.ts, with the alias and setupFiles paths rebased from
  './tools/...' and './__mocks__/...' to './...' and '../__mocks__/...'
  respectively (it now lives one directory deeper).
- vitest.config.ts: collapsed to 're-export the shared config' so
  running 'vitest' at the repo root keeps working.
- packages/*/vitest.config.ts: nine new thin shims, one per app, that
  do mergeConfig(sharedConfig, {}). Each app now has a discoverable
  config when vitest is run from inside the app directory.
- packages/*/package.json: drop the '-T' flag from the test scripts
  ('yarn run -T vitest run --passWithNoTests' -> 'vitest run
  --passWithNoTests'); add 'vitest' as a per-app devDependency so the
  binary resolves under yarn 4 without the workspace-root hop.

692 tests pass across the 8 packages that have tests (esm-home-app has
none); 2 skips unchanged.
