+++ packages/esm-patient-registration-app/vitest.config.ts
@@ -1,0 +1,4 @@
+import { mergeConfig } from 'vitest/config';
+import sharedConfig from '../../tools/vitest.shared';
+
+export default mergeConfig(sharedConfig, {});
